moving share.NewModule to constructors
