+++ nodebuilder/share/constructors.go
@@ -1,16 +1,19 @@
 package share
 
 import (
-	"go.uber.org/fx"
+	"context"
 
+	"github.com/ipfs/go-blockservice"
 	"github.com/ipfs/go-datastore"
 	"github.com/libp2p/go-libp2p-core/host"
 	"github.com/libp2p/go-libp2p-core/routing"
 	routingdisc "github.com/libp2p/go-libp2p/p2p/discovery/routing"
+	"go.uber.org/fx"
 
 	"github.com/celestiaorg/celestia-node/share"
 	"github.com/celestiaorg/celestia-node/share/availability/cache"
 	"github.com/celestiaorg/celestia-node/share/availability/discovery"
+	"github.com/celestiaorg/celestia-node/share/service"
 )
 
 func Discovery(cfg Config) func(routing.ContentRouting, host.Host) *discovery.Discovery {
@@ -35,4 +38,17 @@ func CacheAvailability[A share.Availability](lc fx.Lifecycle, ds datastore.Batch
 		OnStop: ca.Close,
 	})
 	return ca
+}
+
+func NewModule(lc fx.Lifecycle, bServ blockservice.BlockService, avail share.Availability) Module {
+	serv := service.NewShareService(bServ, avail)
+	lc.Append(fx.Hook{
+		OnStart: func(ctx context.Context) error {
+			return serv.Start(ctx)
+		},
+		OnStop: func(ctx context.Context) error {
+			return serv.Stop(ctx)
+		},
+	})
+	return serv
 }
